+++ src/main/java/cn/projectan/strix/model/dict/system/PayOrderStatus.java
@@ -14,18 +14,22 @@ import cn.projectan.strix.model.dict.base.BaseDict;
 public class PayOrderStatus implements BaseDict {
 
     @DictData(label = "未支付", sort = 1, style = DictDataStyle.DEFAULT)
-    public static final int UNPAID = 1;
+    public static final
+    short UNPAID = 1;
 
     @DictData(label = "已支付", sort = 2, style = DictDataStyle.SUCCESS)
-    public static final int PAID = 2;
+    public static final
+    short PAID = 2;
 
     @DictData(label = "已退款", sort = 3, style = DictDataStyle.ERROR)
-    public static final int REFUNDED = 3;
+    public static final
+    short REFUNDED = 3;
 
     @DictData(label = "超时未支付", sort = 99, style = DictDataStyle.WARNING)
-    public static final int EXPIRED = 99;
+    public static final
+    short EXPIRED = 99;
 
-    public static boolean valid(byte value) {
+    public static boolean valid(short value) {
         return value == UNPAID || value == PAID || value == REFUNDED || value == EXPIRED;
     }
