+++ src/main/java/cn/projectan/strix/model/other/system/module/sms/StrixSmsTemplate.java
@@ -19,9 +19,9 @@ public class StrixSmsTemplate {
 
     private String name;
 
-    private Integer type;
+    private Short type;
 
-    private Integer status;
+    private Short status;
 
     private String content;
